+++ src/form-decimal-field.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import {useFieldProps} from './form-context';
 import {useTheme} from './hooks';
-
-import type {CommonFormFieldProps} from './form';
-import {TextFieldBase} from './text-field-base';
+import TextFieldBase from './text-field-base';
 import {Locale} from './utils/locale';
+
+import type {CommonFormFieldProps} from './text-field-base';
 
 const getLocalDecimalChar = (locale: Locale): string => {
     try {
@@ -111,3 +111,5 @@ export const FormDecimalField: React.FC<FormDecimalFieldProps> = ({
 
     return <TextFieldBase {...rest} {...fieldProps} inputComponent={DecimalInput} />;
 };
+
+export default FormDecimalField;