+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+# Unreleased
+
+- Fix Switch component when used in uncontrolled mode
+
 ## v.5.0.0 - 2020-07-29
 
 ### Breaking changes