+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## Next
+
+- **Internal**: Changed codemod `import-type.js`, now it transforms to import type (flow) everything imported
+  as a type in typescript.
+
 ## v.5.0.0 - 2020-07-29
 
 ### Breaking changes