+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix Switch component when used in uncontrolled mode
 - **Internal**: Changed codemod `import-type.js`, now it transforms to import type (flow) everything imported
   as a type in typescript.
+- **Internal**: Code reorganization
 
 ## v.5.0.0 - 2020-07-29